+++ recipes/setup_envars.rb
@@ -15,11 +15,24 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-ruby_block 'Configure environment variable: PATH' do
+# PATH
+# PATH envar must include a set of directories, both within the recipe context and the resulting system
+path_required_directories = %w[/usr/local/sbin /usr/local/bin /sbin /bin /usr/sbin /usr/bin /opt/aws/bin]
+
+# This block configures PATH for the recipes context
+ruby_block 'Configure environment variable for recipes context: PATH' do
   block do
-    directories = %w[/usr/local/sbin /usr/local/bin /sbin /bin /usr/sbin /usr/bin /opt/aws/bin]
-    directories.each do |directory|
+    path_required_directories.each do |directory|
       ENV['PATH'] = "#{ENV['PATH']}:#{directory}" unless ":#{ENV['PATH']}:".include?(":#{directory}:")
     end
   end
+end
+
+# This block configures PATH for the system
+template '/etc/profile.d/path.sh' do
+  source 'profile/path.sh.erb'
+  owner 'root'
+  group 'root'
+  mode '0755'
+  variables(path_required_directories: path_required_directories)
 end