+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 **CHANGES**
 
 - Drop support for SGE and Torque schedulers.
-- Drop support for CentOS8.  
+- Drop support for CentOS8.
 - Remove nodewatcher, sqswatcher, jobwatcher related code.
 - Remove Ganglia support.
 - Install ParallelCluster AWS Batch CLI at AMI build time.
@@ -18,6 +18,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Add explicit assignment of names, uids, gids for slurm, munge and dcvextauth users.
 - Remove packer.
 - Restrict access to IMDS to root and cluster admin users, only.
+- Make PATH include required directories for every user and recipes context.
 
 
 2.x.x